chore(test): set up client side test
-create snapshot test

[Finishes #155269192]
