+++ client/__tests__/actions/recipe.test.js
@@ -177,52 +177,6 @@ describe('All actions', () => {
     });
   });
 
-  describe('Sign up a user', () => {
-    test('Should dispatch signup action when signup action is called', async () => {
-      moxios.wait(() => {
-        const getRecipesRequest = moxios.requests.mostRecent();
-        getRecipesRequest.respondWith({
-          status: 201,
-          response: userData
-        });
-      });
-
-      const store = mockStore({});
-      const { user, token } = userData;
-      const expectedAction = {
-        type: APPCONSTANT.SIGN_UP,
-        payload: {
-          user, token
-        }
-      };
-      await store.dispatch(signup());
-      expect(store.getActions()[0]).toEqual(expectedAction);
-    });
-  });
-
-  describe('Sign in a user', () => {
-    test('Should dispatch signin action when signin action is called', async () => {
-      moxios.wait(() => {
-        const getRecipesRequest = moxios.requests.mostRecent();
-        getRecipesRequest.respondWith({
-          status: 201,
-          response: userData
-        });
-      });
-
-      const store = mockStore({});
-      const { user, token } = userData;
-      const expectedAction = {
-        type: APPCONSTANT.SIGN_IN,
-        payload: {
-          user, token
-        }
-      };
-      await store.dispatch(signin());
-      expect(store.getActions()[0]).toEqual(expectedAction);
-    });
-  });
-
   describe('Delete a recipe', () => {
     test('Should dispatch deleteRecipe action when deleteRecipe action is called', async () => {
       moxios.wait(() => {
@@ -274,141 +228,6 @@ describe('All actions', () => {
         ingredients: 'ingredeints'
       };
       await store.dispatch(editRecipe(newRecipe, 1));
-      expect(store.getActions()[0]).toEqual(expectedAction);
-    });
-  });
-
-  describe('Edit user profile', () => {
-    test('Should dispatch editUserProfile action when editUserProfile action is called', async () => {
-      moxios.wait(() => {
-        const getRecipesRequest = moxios.requests.mostRecent();
-        getRecipesRequest.respondWith({
-          status: 200,
-          response: updatedUserData
-        });
-      });
-
-      localStorage.userData = JSON.stringify({
-        token: 'some_token'
-      });
-
-      const store = mockStore({});
-      const expectedAction = {
-        type: APPCONSTANT.EDIT_USER_PROFILE,
-        payload: updatedUserData
-      };
-      // console.log(expectedAction)
-      const newUser = {
-        firstName: 'Ayobami'
-      };
-      await store.dispatch(editUserProfile(newUser, 1));
-      expect(store.getActions()[0]).toEqual(expectedAction);
-    });
-  });
-
-  describe('Edit user password', () => {
-    test('Should dispatch editUserPassword action when editUserPassword action is called', async () => {
-      moxios.wait(() => {
-        const getRecipesRequest = moxios.requests.mostRecent();
-        getRecipesRequest.respondWith({
-          status: 200,
-          response: updateUserPassword
-        });
-      });
-
-      localStorage.userData = JSON.stringify({
-        token: 'some_token'
-      });
-
-      const store = mockStore({});
-      const expectedAction = {
-        type: APPCONSTANT.EDIT_USER_PASSWORD,
-        payload: updateUserPassword
-      };
-      const newUser = {
-        firstName: 'Ayobami'
-      };
-      await store.dispatch(editUserPassword(newUser, 1));
-      expect(store.getActions()[0]).toEqual(expectedAction);
-    });
-  });
-
-  describe('Add revieww', () => {
-    test('Should dispatch addReview action when addReview action is called', async () => {
-      moxios.wait(() => {
-        const getRecipesRequest = moxios.requests.mostRecent();
-        getRecipesRequest.respondWith({
-          status: 200,
-          response: review
-        });
-      });
-
-      localStorage.userData = JSON.stringify({
-        token: 'some_token'
-      });
-
-      const store = mockStore({});
-      const expectedAction = {
-        type: APPCONSTANT.ADD_REVIEW,
-        payload: review
-      };
-
-      const userReview = {
-        review: 'some review',
-      };
-
-
-      await store.dispatch(addReview(userReview, 1));
-      expect(store.getActions()[0]).toEqual(expectedAction);
-    });
-  });
-
-  describe('Get recipe review', () => {
-    test('Should dispatch getRecipeReviews action when getRecipeReviews action is called', async () => {
-      moxios.wait(() => {
-        const getRecipesRequest = moxios.requests.mostRecent();
-        getRecipesRequest.respondWith({
-          status: 200,
-          response: getReview
-        });
-      });
-
-      localStorage.userData = JSON.stringify({
-        token: 'some_token'
-      });
-
-      const store = mockStore({});
-      const expectedAction = {
-        type: APPCONSTANT.GET_RECIPES_REVIEWS,
-        payload: getReview.data
-      };
-
-      await store.dispatch(getRecipeReviews(1));
-      expect(store.getActions()[0]).toEqual(expectedAction);
-    });
-  });
-
-  describe('Get all users', () => {
-    test('Should dispatch getUsers action when getUsers action is called', async () => {
-      moxios.wait(() => {
-        const getRecipesRequest = moxios.requests.mostRecent();
-        getRecipesRequest.respondWith({
-          status: 200,
-          response: allUsers
-        });
-      });
-
-      localStorage.userData = JSON.stringify({
-        token: 'some_token'
-      });
-
-      const store = mockStore({});
-      const expectedAction = {
-        type: APPCONSTANT.FIND_USERS,
-        payload: allUsers
-      };
-
-      await store.dispatch(getUsers());
       expect(store.getActions()[0]).toEqual(expectedAction);
     });
   });
