+++ client/__tests__/setupTests.js
@@ -1,0 +1,4 @@
+import { configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16'; 
+
+configure({ adapter: new Adapter() });
